requirements.txt
Django>=3.2,<4.0
gunicorn
psycopg2-binary
pillow
